requirements.txt
# Python requirements
paho-mqtt
aiohttp
